+++ client/src/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import "./css/App.css";
-import Navigation from "./components/Navigation";
+import Navigation from "./components/Navigation/Navigation";
 
 class App extends Component {
   render() {
