--- client/src/components/Error.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Error = () => {
-  return (
-    <div>
-      <p>Error: Page does not exist!</p>
-    </div>
-  );
-};
-
-export default Error;
