--- client/src/css/Instructions.css
@@ -1,22 +1,0 @@
-.Instructions {
-  display: flex;
-  width: 100%;
-  height: 100%;
-  background: #6e828b83;
-  min-height: 100vh;
-  flex-direction: column;
-  flex-wrap: wrap;
-}
-
-p {
-  padding: 20px;
-}
-
-h1 {
-  text-align: center;
-}
-
-.container {
-  margin-left: 20px;
-  margin-right: 20px;
-}
