--- client/src/css/Signup.css
@@ -1,46 +1,0 @@
-.signupContainer {
-  text-align: center;
-  display: flex;
-  width: 100%;
-  height: 100%;
-  background: #6e828b83;
-  min-height: 100vh;
-}
-
-.signupPage {
-  text-align: center;
-  width: 50vh;
-  margin: 100px auto;
-  color: white;
-  background: black;
-  padding: 10px;
-  line-height: 15px;
-  border-radius: 15px;
-  min-height: 400px;
-  max-height: 400px;
-}
-
-.signupPage label {
-  display: inline-block;
-  width: 50%;
-  text-align: left;
-  margin: 0 0;
-  padding: 5px;
-}
-
-input {
-  width: 80%;
-  padding: 12px;
-  border: none;
-  border-radius: 4px;
-  margin: 5px 0;
-  opacity: 0.85;
-  display: inline-block;
-  font-size: 17px;
-  line-height: 20px;
-  text-decoration: none; /* remove underline from anchors */
-}
-
-input:hover {
-  opacity: 1;
-}
