--- client/src/css/index.css
@@ -1,9 +1,0 @@
-.App {
-  display: flex;
-  width: 100%;
-  height: 100%;
-  background: #6e828b83;
-  min-height: 100vh;
-  flex-direction: column;
-  flex-wrap: wrap;
-}
